refactor: Player 카드 뽑기 테스트를 Challenger 테스트로 이동
